requirements.txt
Flask==2.2.3
gdown==4.7.1
matlab==0.1
matplotlib==3.6.3
multimatch_gaze==0.1.3
numpy==1.19.5
pandas==1.4.4
PyYAML==6.0
scipy==1.10.0
setuptools==56.0.0
pytest==7.2.2

# Image processing
opencv_python==4.5.1.48
Pillow==9.5.0
scikit_image==0.19.3

# Displaying data nicely in the terminal
colorlog==6.7.0
tqdm==4.64.1

# Deep libraries
Theano==1.0.5
torch==2.0.1
torchvision==0.15.2
keras==2.12.0

# Used in metrics
dtaidistance==2.3.10
fastdtw==0.3.4
